pyproject.toml
[tool.poetry]
name = "vandr-fastapi-db-sqs-template"
version = "0.1.0"
description = ""
authors = ["Viktor Andriichuk <v.andriichuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.13.1"
aiosqlite = "^0.20.0"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.27"
uvicorn = "^0.27.1"
aioboto3 = "^12.3.0"
greenlet = "^3.0.3"
ruff = "^0.2.2"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
